pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-wildfire_env"
version = "0.1.0"
description = "Wildfire Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0"
]

[project.optional-dependencies]
cnn = ["tensorflow-cpu>=2.13.0", "opencv-python-headless>=4.8.0"]

[project.scripts]
server = "wildfire_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["wildfire_env", "wildfire_env.server"]
package-dir = { "wildfire_env" = ".", "wildfire_env.server" = "server" }